リファクタリング: ストレージユーティリティを型分離した専用のライブラリディレクトリに移行する
