+++ features/options/SubCategoryKeywordManager.tsx
@@ -1,6 +1,8 @@
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
+import { setCategoryKeywords } from '@/lib/storage'
+import type { TabGroup } from '@/types/storage'
 import { Check, Plus, X } from 'lucide-react'
 import { useRef, useState } from 'react'
