+++ features/saved-tabs/components/ProjectUrlItem.tsx
@@ -1,6 +1,6 @@
 import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
-import type { CustomProject, UserSettings } from '@/utils/storage'
+import type { CustomProject, UserSettings } from '@/types/storage'
 import { useSortable } from '@dnd-kit/sortable'
 import { CSS } from '@dnd-kit/utilities'
 import { ChevronRight, GripVertical, X } from 'lucide-react'
