+++ types/index.ts
@@ -1,0 +1,5 @@
+// ストレージ関連の型定義
+export * from './storage'
+
+// saved-tabs関連の型定義（既存のファイルがあれば再エクスポート）
+export * from './saved-tabs'
