+++ types/storage.ts
@@ -1,0 +1,83 @@
+// 親カテゴリのインターフェース
+export interface ParentCategory {
+  id: string
+  name: string
+  domains: string[] // このカテゴリに属するドメインIDのリスト
+  domainNames: string[] // このカテゴリに属するドメイン名のリスト (新規追加)
+}
+
+// 子カテゴリのキーワード設定のインターフェース
+export interface SubCategoryKeyword {
+  categoryName: string // カテゴリ名
+  keywords: string[] // 関連キーワードリスト
+}
+
+export interface TabGroup {
+  id: string
+  domain: string
+  parentCategoryId?: string // 親カテゴリのID
+  urls: {
+    url: string
+    title: string
+    subCategory?: string // 子カテゴリ名
+    savedAt?: number // 個別のURL保存時刻を追加
+  }[]
+  subCategories?: string[] // このドメインで利用可能な子カテゴリのリスト
+  categoryKeywords?: SubCategoryKeyword[] // 子カテゴリのキーワード設定
+  subCategoryOrder?: string[] // 子カテゴリの表示順序
+  subCategoryOrderWithUncategorized?: string[] // 未分類カテゴリを含む全カテゴリの表示順序
+  savedAt?: number // グループ全体の保存時刻を追加
+}
+
+export interface UserSettings {
+  removeTabAfterOpen: boolean
+  excludePatterns: string[]
+  enableCategories: boolean // カテゴリ機能の有効/無効
+  autoDeletePeriod?: string // never, 1hour, 1day, 7days, 14days, 30days, 180days, 365days
+  showSavedTime: boolean // 保存日時を表示するかどうか
+  clickBehavior:
+    | 'saveCurrentTab'
+    | 'saveWindowTabs'
+    | 'saveSameDomainTabs'
+    | 'saveAllWindowsTabs' // 新しいオプション: クリック時の挙動
+  excludePinnedTabs: boolean // 固定タブ（ピン留め）を除外するかどうか
+  openUrlInBackground: boolean // URLクリック時に別タブで開くかどうか
+  openAllInNewWindow: boolean // 「すべてのタブを開く」を別ウィンドウで開くかどうか
+  confirmDeleteAll: boolean // すべて削除前に確認するかどうか
+  confirmDeleteEach: boolean // URL削除前に個別確認するかどうか
+  colors?: Record<string, string> // ユーザー設定: カラー設定まとめ
+}
+
+// ドメイン別のカテゴリ設定を保存するためのインターフェース
+export interface DomainCategorySettings {
+  domain: string // ドメイン
+  subCategories: string[] // このドメインで設定された子カテゴリリスト
+  categoryKeywords: SubCategoryKeyword[] // カテゴリキーワード設定
+}
+
+// ドメインと親カテゴリのマッピングを保存するインターフェース
+export interface DomainParentCategoryMapping {
+  domain: string // ドメイン（URL）
+  categoryId: string // 親カテゴリID
+}
+
+// カスタムプロジェクト（PJ単位）のデータ構造
+export interface CustomProject {
+  id: string
+  name: string
+  description?: string
+  urls: {
+    url: string
+    title: string
+    notes?: string
+    savedAt?: number // 個別のURL保存時刻
+    category?: string // プロジェクト内でのカテゴリ分類
+  }[]
+  categories: string[] // このプロジェクトで利用可能なカテゴリリスト
+  categoryOrder?: string[] // カテゴリの表示順序
+  createdAt: number
+  updatedAt: number
+}
+
+// ビューモード（表示モード）の型定義
+export type ViewMode = 'domain' | 'custom'
